AutoMapper

+++ BookStore/BookStore/Controllers/AuthorController.cs
@@ -1,6 +1,8 @@
 ﻿using BookStore.BL.Interfaces;
 using BookStore.Models.Models;
+using BookStore.Models.Requests;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
 
 namespace BookStore.Controllers
 {
@@ -16,30 +18,72 @@ namespace BookStore.Controllers
             _authorService = authorService;
         }
 
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [HttpGet(nameof(GetAllAuthors))]
-        public IEnumerable<Author> GetAllAuthors()
+        public IActionResult GetAllAuthors()
         {
-            return _authorService.GetAllAuthors();
+            return Ok(_authorService.GetAllAuthors());
         }
+
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpGet(nameof(GetById))]
-        public Author? GetById(int id)
+        public IActionResult GetById(int id)
         {
-            return _authorService.GetById(id);
+            if (id <= 0)
+            {
+                _logger.LogInformation("Id must be greater than 0");
+                return BadRequest($"Parameter id: {id} must be greater than 0");
+            }
+
+            var result = _authorService.GetById(id);
+
+            if (result == null) return NotFound(id);
+
+            return Ok(result);
         }
+
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPost(nameof(AddMethod))]
-        public void AddMethod(Author author)
+        public IActionResult AddMethod([FromBody] AddAuthorRequest authorRequest)
         {
-            _authorService.AddAuthor(author);
+            //if (authorRequest == null) return BadRequest(authorRequest);
+            //var authorExist = _authorService.GetAuthorByName(authorRequest.Name);
+            //if (authorExist != null) return BadRequest("Author Already Exist!");
+
+            var result = _authorService.AddAuthor(authorRequest);
+
+            if (result.HttpStatusCode == HttpStatusCode.BadRequest)
+                return BadRequest(result);
+
+            return Ok(result);
         }
+
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPut(nameof(UpdateMethod))]
-        public void UpdateMethod(Author author)
+        public IActionResult UpdateMethod([FromBody] UpdateAuthorRequest authorRequest)
         {
-            _authorService.UpdateAuthor(author);
+            var result = _authorService.UpdateAuthor(authorRequest);
+            if (result.HttpStatusCode == HttpStatusCode.BadRequest)
+                return BadRequest(result);
+
+            return Ok(result);
         }
+
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpDelete(nameof(DeleteMethod))]
-        public Author? DeleteMethod(int id)
+        public IActionResult DeleteMethod(int id)
         {
-            return _authorService.DeleteAuthorById(id);
+            if (id > 0 && _authorService.GetById != null)
+            {
+                _authorService.DeleteAuthorById(id);
+                return Ok($"Author with id {id} is successfully deleted");
+            }
+            return BadRequest($"Author with id {id} is not deleted");
         }
     }
 }
